server/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
transformers
numpy
Pillow
httpx
pydantic-settings
